backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cryptography==45.0.2
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
Faker==37.3.0
fastapi==0.115.12
fastapi-cache2==0.2.2
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
limits==5.2.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
passlib==1.7.4
pendulum==3.1.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
redis==4.6.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
setuptools==80.7.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
psycopg2-binary
huggingface_hub[hf_xet]

